Refactor WebGL render tests
