+++ test/spec/ol/render/webgl/index.test.js
@@ -1,0 +1,1 @@
+goog.provide('ol.test.render.webgl.Replay');
